Continue with not main files
